+++ Tests/GeneratedTests/GeneratedTests.swift
@@ -726,6 +726,12 @@ class NoGroupingExtensionRuleGeneratedTests: XCTestCase {
     }
 }
 
+class NoMagicNumbersRuleGeneratedTests: XCTestCase {
+    func testWithDefaultConfiguration() {
+        verifyRule(NoMagicNumbersRule.description)
+    }
+}
+
 class NoSpaceInMethodCallRuleGeneratedTests: XCTestCase {
     func testWithDefaultConfiguration() {
         verifyRule(NoSpaceInMethodCallRule.description)